Simplify all vehicle API calls by accepting a VIN directly instead of a Vehicle instance.
This removes the need to have an instance of a Vehicle in order to call the APIs, and is particularly handy in cases where you may want to make network requests but you only have a VIN available (e.g. when being used as a key in a dictionary).

This is a breaking change. Fixing it is often just a matter of replacing `vehicle: vehicle` with `vin: vehicle.vin` or `vin: vin` if you already have the vin as a variable.