Route initial data load through the refresh handler.
This will allow the refresh UI to reflect the status of the refresh on app launch.

ModelStore now also provides a way to always get access to a Subject for each given network stream. This ensures that even if the UI isn't active that there's still some Subject storing state from the network requests.